+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the nginx cookbook.
 ## Unreleased
 
 - Fix installation on EL9 platforms
+- Fix installation on newer Amazon Linux platforms
 
 ## 12.1.7 - *2023-04-04*